requirements.txt
numpy>=1.26.0,<2.0
opencv-python>=4.8.0
Pillow>=10.0.0
pydantic>=2.0
tqdm
datasets>=2.14,<3.0
huggingface-hub>=0.23.0
boto3>=1.34.0
